--- common/azure_auth.go
@@ -1,50 +1,0 @@
-package common
-
-import (
-	"fmt"
-	"log"
-	"net/http"
-	"strings"
-
-	"github.com/golang-jwt/jwt/v4"
-)
-
-func (aa *DatabricksClient) GetAzureJwtProperty(key string) (any, error) {
-	if !aa.IsAzure() {
-		return "", fmt.Errorf("can't get Azure JWT token in non-Azure environment")
-	}
-	if key == "tid" && aa.Config.AzureTenantID != "" {
-		return aa.Config.AzureTenantID, nil
-	}
-	request, err := http.NewRequest("GET", aa.Config.Host, nil)
-	if err != nil {
-		return nil, err
-	}
-	err = aa.Config.Authenticate(request)
-	if err != nil {
-		return nil, err
-	}
-	header := request.Header.Get("Authorization")
-	var stoken string
-	if len(header) > 0 && strings.HasPrefix(string(header), "Bearer ") {
-		log.Printf("[DEBUG] Got Bearer token")
-		stoken = strings.TrimSpace(strings.TrimPrefix(string(header), "Bearer "))
-	}
-	if stoken == "" {
-		return nil, fmt.Errorf("can't obtain Azure JWT token")
-	}
-	if strings.HasPrefix(stoken, "dapi") {
-		return nil, fmt.Errorf("can't use Databricks PAT")
-	}
-	parser := jwt.Parser{SkipClaimsValidation: true}
-	token, _, err := parser.ParseUnverified(stoken, jwt.MapClaims{})
-	if err != nil {
-		return nil, err
-	}
-	claims := token.Claims.(jwt.MapClaims)
-	v, ok := claims[key]
-	if !ok {
-		return nil, fmt.Errorf("can't find field '%s' in parsed JWT", key)
-	}
-	return v, nil
-}
